Refactoring chart components
